Implement placeholders in Compose
