+++ media-placeholders/build.gradle
@@ -32,6 +32,12 @@ android {
             includeAndroidResources = true
         }
     }
+    buildFeatures {
+        compose true
+    }
+    composeOptions {
+        kotlinCompilerExtensionVersion = "1.5.14"
+    }
 }
 
 dependencies {
@@ -41,6 +47,10 @@ dependencies {
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$kotlinCoroutinesVersion"
     testImplementation "junit:junit:$jUnitVersion"
     testImplementation "org.robolectric:robolectric:$robolectricVersion"
+    implementation 'androidx.compose.material3:material3:1.3.1'
+    implementation 'androidx.compose.foundation:foundation:1.7.5'
+    implementation 'androidx.compose.ui:ui:1.7.5'
+    implementation 'androidx.compose.ui:ui-tooling-preview:1.7.5'
 }
 
 dependencyAnalysis {
