+++ media-placeholders/src/main/java/org/wordpress/aztec/placeholders/ComposePlaceholderAdapter.kt
@@ -1,0 +1,14 @@
+package org.wordpress.aztec.placeholders
+
+import androidx.compose.runtime.Composable
+import org.wordpress.aztec.AztecAttributes
+
+interface ComposePlaceholderAdapter : PlaceholderManager.PlaceholderAdapter {
+    @Composable
+    fun Placeholder(
+        placeholderUuid: String,
+        attrs: AztecAttributes,
+        width: Int,
+        height: Int,
+    ) {}
+}
